Chore: Dehydrate login methods from rocketchat.js (#3759)
* dehydrate login methods from rocketchat.js
