+++ app/definitions/ICredentials.ts
@@ -1,0 +1,15 @@
+export interface ICredentials {
+	user?: string;
+	password?: string;
+	username?: string;
+	ldapPass?: string;
+	ldap?: boolean;
+	ldapOptions?: object;
+	crowdPassword?: string;
+	crowd?: boolean;
+	code?: string;
+	totp?: {
+		login: ICredentials;
+		code: string;
+	};
+}
