+++ app/definitions/index.ts
@@ -24,6 +24,7 @@ export * from './IServerHistory';
 export * from './IRocketChat';
 export * from './ICertificate';
 export * from './IUrl';
+export * from './ICredentials';
 export * from './ISearch';
 
 export interface IBaseScreen<T extends Record<string, object | undefined>, S extends string> {
